Move non-command transaction functions to another module
