+++ cardano-cli/cardano-cli.cabal
@@ -156,6 +156,7 @@ library
                         Cardano.CLI.Types.Key
                         Cardano.CLI.Types.Key.VerificationKey
                         Cardano.CLI.Types.Output
+                        Cardano.CLI.Types.Transaction
                         Cardano.CLI.Types.TxFeature
 
   other-modules:        Paths_cardano_cli
